// src/app/models/van-ban-den/tim-kiem-theo-so-van-ban.ts
export interface TraCuuNangCao {}

export interface TimKiemDanhSachTheoSoVanBan {
    keyWord?: string;
    soVanBanId?: number;
    vanBanId?: number;
    donViId?: number;
    loaiVanBanId?: number;
    ngayGuiTuNgay?: string;
    ngayGuiDenNgay?: string;
    banHanhTuNgay?: string;
    banHanhDenNgay?: string;
    nam?: number;
    thang?: number;
    trangThai?: number;
    soKyHieu?: string;
    soDi?: number;
    trichYeu?: string;
    timChinhXac?: number;
    cQBHId?: number;
}